+++ package.json
@@ -15,13 +15,13 @@
     "op-indexer": "source .env.local && export DATABASE_URL && export NEXT_PUBLIC_L1_RPC_URL && export NEXT_PUBLIC_L2_RPC_URL && export L1_RPC_WSS && export L2_RPC_WSS && bun op-indexer/index.ts"
   },
   "dependencies": {
-    "@prisma/client": "^5.15.1",
+    "@prisma/client": "^5.16.0",
     "@radix-ui/react-avatar": "^1.1.0",
     "@radix-ui/react-dialog": "^1.1.0",
     "@radix-ui/react-dropdown-menu": "^2.1.0",
     "@radix-ui/react-progress": "^1.1.0",
     "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/react-tooltip": "^1.1.0",
+    "@radix-ui/react-tooltip": "^1.1.1",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "date-fns": "^3.6.0",
@@ -35,10 +35,10 @@
     "recharts": "^2.12.7",
     "tailwind-merge": "^2.3.0",
     "tailwindcss-animate": "^1.0.7",
-    "viem": "^2.16.0"
+    "viem": "^2.16.3"
   },
   "devDependencies": {
-    "@types/node": "^20.14.7",
+    "@types/node": "^20.14.9",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "@types/react-syntax-highlighter": "^15.5.13",
@@ -48,7 +48,7 @@
     "postcss": "^8.4.38",
     "prettier": "^3.3.2",
     "prettier-plugin-tailwindcss": "^0.6.5",
-    "prisma": "^5.15.1",
+    "prisma": "^5.16.0",
     "tailwindcss": "^3.4.4",
     "typescript": "^5.5.2"
   }
