+++ src/components/pages/tx/index.tsx
@@ -2,12 +2,16 @@ import { Hash } from "viem";
 import { l2PublicClient } from "@/lib/chains";
 import TransactionDetails from "@/components/pages/tx/transaction-details";
 import { parseTokenTransfers } from "@/lib/utils";
+import { notFound } from "next/navigation";
 
 const Tx = async ({ hash }: { hash: Hash }) => {
   const [transaction, receipt] = await Promise.all([
     l2PublicClient.getTransaction({ hash }),
     l2PublicClient.getTransactionReceipt({ hash }),
   ]);
+  if (!transaction || !receipt) {
+    notFound();
+  }
   const tokenTransfers = await parseTokenTransfers(receipt.logs);
   return (
     <main className="flex flex-1 flex-col gap-4 p-4 md:gap-4 md:p-4">
